Move test_rank to directories
